modules renamed to lowercase
