+++ inflector.py
@@ -6,8 +6,8 @@
 # See the end of this file for the free software, open source license (BSD-style).
 
 import re
-from Rules.English import English
-from Rules.Spanish import Spanish
+from rules.english import English
+from rules.spanish import Spanish
 
 class Inflector :
     """
