+++ rules/base.py
