+++ rules/spanish.py
@@ -9,7 +9,7 @@
 # See the end of this file for the free software, open source license (BSD-style).
 
 import re
-from Base import Base
+from base import Base
 
 class Spanish (Base):
     '''
